* Works for BrowserStack - #Closes 9 * Needs investigation for Sauce Labs * Renaming packages to make it more "generic" and remove the vanity mention
