+++ src/main/java/qa/free/tools/selenium/webdriver/factory/drivers/options/DefaultOptions.java
@@ -1,0 +1,22 @@
+package qa.free.tools.selenium.webdriver.factory.drivers.options;
+
+import org.aeonbits.owner.ConfigFactory;
+import org.openqa.selenium.MutableCapabilities;
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
+
+import lombok.AccessLevel;
+import lombok.Getter;
+import qa.free.tools.selenium.webdriver.properties.WebDriverProperties;
+
+public abstract class DefaultOptions {
+
+	@Getter(AccessLevel.PROTECTED)
+	WebDriverProperties webDriverProperties = ConfigFactory.create(WebDriverProperties.class);
+	
+	protected final Logger logger = LoggerFactory.getLogger(this.getClass());
+	
+	public abstract MutableCapabilities getOptions();
+	
+	
+}
